refactor: change name of index.tsx to the corresponding components
