+++ src/App.tsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { GlobalStyle } from './App.styled';
-import NavBar from './components/NavBar';
-import NotFound from './components/NotFound';
-import Albums from './pages/Albums';
-import Photos from './pages/Photos';
-import Profile from './pages/Profile';
+import NavBar from './components/NavBar/NavBar';
+import NotFound from './components/NotFound/NotFound';
+import Albums from './pages/Albums/Albums';
+import Photos from './pages/Photos/Photos';
+import Profile from './pages/Profile/Profile';
 
 const App = () => (
   <BrowserRouter>
