+++ src/components/Button/Button.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BtnWrapper } from './Button.styled';
-import { ButtonConfig } from './Button';
+import { ButtonConfig } from './Button.d';
 
 const Button = ({ children, color, link, type }: ButtonConfig) => {
   const btn: JSX.Element = (
