+++ src/components/Card/Card.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CardWrapper, TitleWrapper } from './Card.styled';
-import { CardConfig } from './Card';
+import { CardConfig } from './Card.d';
 
 const Card = ({ url, title, col, id, handleZoom }: CardConfig) => (
   <CardWrapper onClick={() => handleZoom(id)} col={col}>
