+++ src/components/NoResult/NoResult.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { MessageWrapper } from './NoResult.styled';
 import WarningIcon from '@material-ui/icons/Warning';
-import { NoResultOptions } from './NoResult';
+import { NoResultOptions } from './NoResult.d';
 
 const NoResult = ({ message }: NoResultOptions) => (
   <MessageWrapper>
